# openapi.yml
openapi: 3.0.3
info:
  title: Tasked - OpenAPI 3.0
  description: |-
    Tasked est une application de timbrage. L'application permet de créer des projets, puis de créer des tâches en "timbrant". 
    L'idée est de l'utiliser au sein d'une entreprise afin que l'employé ait simplement à passer son téléphone sur un lecteur NFC pour envoyer une requête de "début de tâche". Idem pour la fin de la journée
    
    Quelques liens utiles:
    - [Cours ArchiOWeb](https://github.com/MediaComem/comem-archioweb)
    - [Projet github "tasked"](https://github.com/emilevl/tasked)
    - [API Specification](https://swagger.io/specification/)
  termsOfService: http://swagger.io/terms/
  version: 1.0.0
...
servers:
  - url: https://tasked.onrender.com/
  - url: http://localhost:3000/
...
paths:
...
  /users:
    get:
      parameters: 
        [
          $ref: '#/components/parameters/pageParam',
          $ref: '#/components/parameters/pageSizeParam'
        ]
      tags:
        - users
      summary: List of users
      description: Get the list of all the users, sorted by firstName ascending
      operationId: getUsers
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
        '404':
          description: Id Not found
      security:
        - user_auth:
            - read:admin
    post:
      tags:
        - users
      summary: Create a user
      description: Create a user with all information
      operationId: createUser
      requestBody:
        description: Send all the user informations
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPost'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Unauthorized
        '404':
          description: Id Not found
      security:
        - user_auth:
            - write:admin
  /users/{idUser}:
    parameters: [$ref: '#/components/parameters/idUserParam']
    patch:
      tags:
        - users
      summary: Edit a user
      description: Edit a user sending only the things to edit as an object
      operationId: editUser
      requestBody:
        description: Send only the information that need to be edited
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPost'
        required: true
      responses:
        '200':
          description: Successful operation - return the whole user informations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Unauthorized
        '404':
          description: Id Not found
      security:
        - user_auth:
            - write:admin
    delete:
      tags:
        - users
      summary: Delete a user
      description: Delete a user using his id.
      operationId: deleteUser
      responses:
        '204':
          description: Successful operation - No content returned
        '401':
          description: Unauthorized
        '404':
          description: Id Not found
      security:
        - user_auth:
            - delete:admin
  /users/register:
    post:
      tags:
        - users
      summary: Create a new account
      description: Create a new account. Anyone can create a new account.
      operationId: register
      requestBody:
        description: Send only what's needed
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterPost'
        required: true
      responses:
        '200':
          description: Successful operation - return the whole user informations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '401':
          description: Unauthorized
        '404':
          description: Not found
      security:
        - user_auth:
            - write:admin
  /auth/login:
    post:
      tags:
        - users
      summary: Authenticate
      description: Authenticate to the server
      operationId: authenticate
      requestBody:
        description: Send the user and password (The user must exist in the collection first)
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserAuthenticate'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserAuthenticated'
        '401':
          description: Unauthorized - Bad login
        '404':
          description: Id Not found
        '500':
          description: Internal Server Error
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          example: 63442662c073e
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        username:
          type: string
          example: johndoe
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/Task"
            description: All the tasks belonging the user
        projects:
          type: array
          items:
            $ref: "#/components/schemas/Project"
            description: All the projects belonging the user
    UserPost:
      type: object
      properties:
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        username:
          type: string
          example: johndoe
        password:
          type: string
        role:
          type: string
          example: user
          default: user
          description: enum ['user', 'editor', 'admin']
      required: 
        - firstName
        - lastName
        - username
        - password
    UserResponse:
      type: object
      properties:
        id:
          type: string
          example: 63442662c073e
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        username:
          type: string
          example: johndoe
        role:
          type: string
          example: user
    RegisterPost:
      type: object
      properties:
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        username:
          type: string
          example: johndoe
        password:
          type: string
      required: 
        - firstName
        - lastName
        - username
        - password
    RegisterResponse:
      type: object
      properties:
        message: 
          type: string
          example: "User successfuly created !"
        id:
          type: string
          example: 63442662c073e
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        username:
          type: string
          example: johndoe
        role:
          type: string
          example: user
    Project:
      type: object
      properties:
        _id:
          type: string
          example: 63442662c073e
        name:
          type: string
          example: My new app
        active:
          type: boolean
        description:
          type: string
        company:
          type: string
        author:
          type: object
          $ref: "#/components/schemas/UserResponse"
          description: Author of the project (user)
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/TaskPost"
            description: All the tasks belonging the project
...
    Task:
      type: object
      properties:
        id:
          type: string
          example: 63442662c073e
        name:
          type: string
          example: My task created
        description:
          type: string
        startDate:
          type: string
          format: date-time
          example: "2022-10-19T13:20:11.686Z"
        endDate:
          format: date-time
          example: "2022-10-19T14:30:11.686Z"
        user:
          type: array
          items:
            $ref: "#/components/schemas/User"
            description: User who created the task
        project:
          type: array
          items:
            $ref: "#/components/schemas/Project"
            description: The project of the task
    TaskPost:
      type: object
      properties:
        name:
          type: string
          example: My task created
        description:
          type: string
        startDate:
          type: string
          format: date-time
          example: "2022-10-19T13:20:11.686Z"
          default: Date of creation
        endDate:
          type: string
          format: date-time
          example: "2022-10-19T14:30:11.686Z"
        project: 
          type: integer
          format: int64
          example: 5604f32fda
          description: id of the project
      required: 
        - name
        - project
...
    UserAuthenticated:
      type: object
      properties:
        token: 
          type: string
          description: token of the authenticated User
    UserAuthenticate:
      type: object
      properties: 
        username:
          type: string
          description: existing user in the collection
        password: 
          type: string
      required: 
        - username
        - password
  parameters:
    pageParam:
      name: page
      in: query
      description: Page index to go (If there are many)
      required: false
      schema:
        type: integer
        format: int32
    pageSizeParam:
      name: pageSize
      in: query
      description: The size of one page (n° of documents)
      required: false
      schema:
        type: integer
        format: int32
...
    idUserParam:
      name: idUser
      in: path
      required: true
      description: Id of the user to retrieve
      schema:
        type: string
...
  securitySchemes:
    user_auth:
      type: http
      description: L'utilisateur doit se connecter pour effectuer des tâches
      scheme: bearer
      bearerFormat: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI2MzRmZDlmZWUyOGMxZDExYzMzMTlmY2UiLCJzY29wZSI6InVzZXIiLCJpYXQiOjE2NjYxODYxMzAsImV4cCI6MTY2Njc5MDkzMH0.06h-L09jS0_rNfQXwCIIcYWBo7zwtSIGz24X0Xmue6A"